admin/requirements.txt
django==4.2.5
flake8==6.1.0
python-dotenv==1.0.0
psycopg2==2.9.9
django-split-settings==1.2.0
uwsgi==2.0.22
